+++ frontend/client/src/components/Navbar/Navbar.jsx
@@ -2,9 +2,12 @@ import React, { useState } from 'react';
 import { FaSearch } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
 import '../../styles/Navbar.css';
+import { useNavigate } from 'react-router-dom';
 
 const Navbar = ({ onSearch }) => {
   const [searchTerm, setSearchTerm] = useState('');
+  const navigate = useNavigate();
+
 
   const handleSearchChange = (event) => {
     setSearchTerm(event.target.value);
@@ -14,11 +17,35 @@ const Navbar = ({ onSearch }) => {
     event.preventDefault();
     onSearch(searchTerm);  // Llama a la función onSearch con el término actual
   };
+
+  const handleLogout = () => {
+    // Clear local storage
+    localStorage.clear();
+    // Redirect to login page
+    window.location.replace("/login");
+    window.location.reload();
+    
+  };
+  const handleRegister = () => {
+    // Redirect to login page
+    navigate("/registro");
+    
+    
+  };
   console.log(localStorage.getItem("usertype"));
   return (
     <nav className="navbar">
       <div className="logo">LP Acquisitions</div>
       <div className="search-bar">
+      <button onClick={handleLogout}>Logout</button>
+      { localStorage.getItem("usertype") === "admin" ?(
+        <>
+        <button onClick={handleRegister}>Crear usuario</button>
+        </>
+      ):(<>
+      </>)
+
+      }
         <form onSubmit={handleSearchSubmit}>
           <input
             type="text"
